+++ app/application/forms/player_name.rb
@@ -6,7 +6,7 @@ module TopPop
   module Forms
     class PlayerName < Dry::Validation::Contract
       # REGEX = %r{.*}.freeze # invalid all the time
-      REGEX = %r{/^\w+\w$/}.freeze # invalid all the time
+      REGEX = %r{^\w+\w$}.freeze # invalid all the time
 
       params do
         required(:player_name).filled(:string)
@@ -14,7 +14,7 @@ module TopPop
 
       rule(:player_name) do
         unless REGEX.match?(value)
-          key.failure('Invalid player name')
+          key.failure('Your username is not valid! Please try a different name!')
         end
       end
     end
