+++ app/presentation/views_html/flash_bar.erb
@@ -1,10 +1,10 @@
 <% if flash[:error] %>
-  <div class="alert alert-danger" id="flash_bar_danger">
+  <div class="mgin11 alert alert-danger" id="flash_bar_danger">
     <%= flash[:error] %>
   </div>
 <% end %>
 <% if flash[:notice] %>
-  <div class="alert alert-success" id="flash_bar_success">
+  <div class="mgin11 alert alert-success" id="flash_bar_success">
     <%= flash[:notice] %>
   </div>
 <% end %>
