+++ app/presentation/views_html/game.erb
@@ -1,5 +1,5 @@
 <div class="container px-lg-5">
   <div class="row">
-    <h2 class="fs-4 fw-bold mgin">The game will be here</h2>
+    <h2 class="fs-4 fw-bold mgin">The game is under development. Please be patient :))</h2>
   </div>
 </div>
