+++ app/presentation/views_html/home.erb
@@ -1,10 +1,11 @@
 <div class="container px-lg-5 main">
+  <div class="mgin11">
+    <h2 class="fs-4 fw-bold">Enter your username</h2>
+    <p>Username should only contains letters, numbers and underscore. For example, you can be "pro_1337" :))</p>        
+  </div>
   <form action="/player" id="player-name-form" method="POST" role="form">
     <div class="row vertical-align query-bar">
-      <div class="col-md-2 col-sm-3">
-        <div class="fs-4 fw-bold">Enter your name</div>        
-      </div>
-      <div class="col-md-8 col-sm-6">
+      <div class="col-md-10 col-sm-3">
         <input class="form-control" id="player-name" name="player_name" type="text" />
       </div>
       <div class="col-md-2 col-sm-3">
