+++ app/presentation/views_html/layout.erb
@@ -21,7 +21,7 @@
     </nav>
 
 
-    <section class="pt-4">
+    <section class="container px-lg-5">
       <%= render('flash_bar') %>
       <%= yield %>
     </section>
